requirements.txt
fastapi
uvicorn
youtube-transcript-api
google-generativeai